requirements.txt
bcrypt
cryptography
customtkinter
mysql-connector-python
openpyxl
pandas
pillow
plyer
psutil
pynput
pystray
python-dotenv
pywin32
reportlab
tendo
tkcalendar
tkinter-tooltip
keyring
winotify
psutil
auto-py-to-exe